+++ CHANGES.rst
@@ -1,6 +1,16 @@
 Release history
 ===============
 
+Version 1.6.1 (2012-08-22)
+--------------------------
+Simplified the installation.
+
+Bugs:
+
+- put pyfire as an optional dependency as it is used only for the campfire backend
+- put PySide as an optional dependency as it is used only for the QT graphical backend
+
+
 Version 1.6.0 (2012-08-16)
 --------------------------
 Bugs: